Replace status code check with is_wp_error
The `wp_safe_remote_get` function returns an `array|WP_Error`. In the previous check, it was too optimistic which was causing errors on a few of the websites I manage. The error was `Uncaught Error: Cannot use object of type WP_Error as array` due to the check of the status code. Performing the `is_wp_error` alleviates this, as anything other than an `WP_Error` is a success in this context anyway.